Adicionando o método Salvar e apagar

+++ Notes/MainPage.xaml.cs
@@ -2,12 +2,12 @@
 {
     public partial class MainPage : ContentPage
     {
-        string caminho = Path.Combine(FileSystem.AppDataDirectory, "arquivo"); 
+        string caminho = Path.Combine(FileSystem.AppDataDirectory, "arquivo");
         public MainPage()
         {
             InitializeComponent();
-            if(File.Exists(caminho))
-            TextEditor.Text = File.ReadAllText(caminho); 
+            if (File.Exists(caminho))
+                TextEditor.Text = File.ReadAllText(caminho);
         }
 
         private void SalvarBtn_Clicked(object sender, EventArgs e)
@@ -18,13 +18,20 @@
         }
         private void ApagarBtn_Clicked(object sender, EventArgs e)
         {
+            TextEditor.Text= string.Empty;
             string conteudo = TextEditor.Text;
-            File.Delete(caminho);
-            DisplayAlert("Arquivo Apagado", "Arquivo Apagado com sucesso", "Ok");
-            TextEditor.Text = string.Empty;
+            if (File.Exists(caminho))
+            {
+                File.Delete(caminho);
+                DisplayAlert("Arquivo Apagado", "Arquivo Apagado com sucesso", "Ok");
+            }
+            else
+            {
+                DisplayAlert("Exclusão","Arquivo não existe","OK");
+            }
+
+
         }
 
-        
     }
-
 }
